refactor : domain, view, controller 로 패키지 분리
