+++ src/main/java/racinggame/domain/racingcar/CarName.java
@@ -1,4 +1,4 @@
-package racinggame.racingcar;
+package racinggame.domain.racingcar;
 
 public class CarName {
     private static final int MAX_NAME_LENGTH = 5;
