+++ src/main/java/racinggame/domain/strategy/RandomMoveStrategy.java
@@ -1,4 +1,4 @@
-package racinggame.racingcar;
+package racinggame.domain.strategy;
 
 import java.util.Random;
