+++ src/main/java/racinggame/view/ResultView.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.view;
 
 import java.util.List;
 import java.util.Map;
