+++ src/test/java/racinggame/RacingCarMovementTest.java
@@ -2,9 +2,9 @@ package racinggame;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racinggame.racingcar.RacingCars;
-import racinggame.racingcar.RandomMoveStrategy;
-import racinggame.racingcar.MoveStrategy;
+import racinggame.domain.racingcar.RacingCars;
+import racinggame.domain.strategy.RandomMoveStrategy;
+import racinggame.domain.strategy.MoveStrategy;
 
 import java.util.Arrays;
 import java.util.List;
