+++ src/test/java/racinggame/RacingGameInputViewTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racinggame.view.ConsoleInputView;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
